requirements.txt
absl-py==2.1.0
aenum==3.1.15
affine==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
arrow==1.3.0
asciitree==0.3.3
async-timeout==4.0.3
attrs==23.2.0
autograd==1.6.2
autograd-gamma==0.5.0
beautifulsoup4==4.12.3
blis==0.7.11
boto3==1.34.67
botocore==1.34.67
bravado==11.0.3
bravado-core==6.1.1
catalogue==2.0.10
cellpose==2.1.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpathlib==0.16.0
cloudpickle==3.0.0
confection==0.1.4
ConfigSpace==0.7.1
contourpy==1.2.0
crc32c==2.3.post0
cucim==23.10.0
cupy-cuda11x==13.0.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.8
dask==2024.2.1
defusedxml==0.7.1
distributed==2024.2.1
efficientnet-pytorch==0.7.1
einops==0.7.0
emcee==3.1.4
fastai==2.7.14
fastcore==1.5.29
fastdownload==0.0.7
fasteners==0.19
fastprogress==1.0.3
fastremap==1.14.1
fastrlock==0.8.2
filelock==3.13.1
fonttools==4.49.0
formulaic==1.0.1
fpdf2==2.7.8
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
gast==0.5.4
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.42
glfw==2.7.0
grpcio==1.62.0
h5py==3.10.0
huggingface-hub==0.21.1
hydra-core==1.3.2
idna==3.6
imagecodecs==2024.1.1
imageio==2.34.0
imgui==2.0.0
importlib-metadata==7.0.1
importlib_resources==6.1.2
interface-meta==1.3.0
isoduration==20.11.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
lifelines==0.28.0
lightly==1.5.0
lightly-utils==0.0.2
lightning-utilities==0.10.1
llvmlite==0.42.0
locket==1.0.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
munch==4.0.0
murmurhash==1.0.10
natsort==8.4.0
neptune==1.9.1
networkx==3.2.1
ninja==1.11.1.1
numba==0.59.0
numcodecs==0.12.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
nystrom-attention==0.0.12
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==1.5.3
parameterized==0.9.0
partd==1.4.1
pillow==10.2.0
preshed==3.0.9
pretrainedmodels==0.7.4
protobuf==3.20.3
psutil==5.9.8
pyarrow==15.0.0
pycparser==2.21
pydantic==1.10.14
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
pynisher==0.6.4
pynndescent==0.5.11
PyOpenGL==3.1.7
pyparsing==3.1.1
pyperclip==1.8.2
pyrfr==0.9.0
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==2.2.0.post0
pytz==2024.1
pyvips==2.2.2
PyYAML==6.0.1
rasterio==1.3.9
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.4.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.18.0
s3transfer==0.10.1
safetensors==0.4.2
saliency==0.2.0
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.11.2
segmentation-models-pytorch==0.3.3
shapely==2.0.3
simplejson==3.19.2
six==1.16.0
slideflow==2.3.1
smac==1.4.0
smart-open==6.4.0
smmap==5.0.1
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
swagger-spec-validator==3.0.3
sympy==1.12
tabulate==0.9.0
tblib==3.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
thinc==8.2.3
threadpoolctl==3.3.0
tifffile==2024.2.12
timm==0.9.2
tokenizers==0.15.2
toolz==0.12.1
torch==2.2.1
torchaudio==2.2.1+cu118
torchmetrics==1.3.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
transformers==4.38.2
triton==2.2.0
typer==0.9.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
umap-learn==0.5.5
uri-template==1.3.0
urllib3==1.26.18
wasabi==1.1.2
weasel==0.3.4
webcolors==1.13
websocket-client==1.7.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
zarr==2.17.0
zict==3.0.0
zipp==3.17.0
